+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - FF-1429 - Updated Credfeto.Enumeration.Source.Generation to 0.0.8.10
+- FF-1429 - Updated FunFair.Test.Common to 6.0.8.2334
 ### Removed
 ### Deployment Changes