+++ CHANGELOG.md
@@ -9,9 +9,9 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Added
 ### Fixed
 ### Changed
-- FF-1429 - Updated Credfeto.Enumeration.Source.Generation to 0.0.8.10
 - FF-1429 - Updated FunFair.Test.Common to 6.0.8.2334
 - FF-1429 - Updated Meziantou.Analyzer to 1.0.750
+- FF-1429 - Updated Credfeto.Enumeration.Source.Generation to 1.0.0.11
 ### Removed
 ### Deployment Changes