+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - FF-1429 - Updated FunFair.Test.Common to 6.0.8.2334
 - FF-1429 - Updated Meziantou.Analyzer to 1.0.750
 - FF-1429 - Updated Credfeto.Enumeration.Source.Generation to 1.0.0.12
+- FF-1429 - Updated SonarAnalyzer.CSharp to 8.49.0.57237
 ### Removed
 ### Deployment Changes