+++ CHANGELOG.md
@@ -9,10 +9,10 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Added
 ### Fixed
 ### Changed
-- FF-1429 - Updated Meziantou.Analyzer to 1.0.750
 - FF-1429 - Updated Credfeto.Enumeration.Source.Generation to 1.0.0.12
 - FF-1429 - Updated SonarAnalyzer.CSharp to 8.49.0.57237
 - FF-1429 - Updated FunFair.Test.Common to 6.0.9.2379
+- FF-1429 - Updated Meziantou.Analyzer to 1.0.752
 ### Removed
 ### Deployment Changes