+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - FF-1429 - Updated xunit.analyzers to 1.1.0
 - FF-1429 - Updated SonarAnalyzer.CSharp to 8.50.0.58025
 - FF-1429 - Updated Microsoft.Extensions to 7.0.1
+- FF-3881 - Updated DotNet SDK to 7.0.101
 ### Removed
 ### Deployment Changes