+++ CHANGELOG.md
@@ -9,7 +9,6 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Added
 ### Fixed
 ### Changed
-- FF-1429 - Updated Credfeto.Enumeration.Source.Generation to 1.0.0.12
 - FF-1429 - Updated Meziantou.Analyzer to 1.0.756
 - FF-1429 - Updated Roslynator.Analyzers to 4.2.0
 - FF-1429 - Updated xunit.analyzers to 1.1.0
@@ -17,6 +16,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - FF-1429 - Updated Microsoft.Extensions to 7.0.1
 - FF-3881 - Updated DotNet SDK to 7.0.101
 - FF-1429 - Updated FunFair.Test.Common to 6.0.10.2422
+- FF-1429 - Updated Credfeto.Enumeration.Source.Generation to 1.0.1.13
 ### Removed
 ### Deployment Changes