+++ CHANGELOG.md
@@ -11,7 +11,6 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Changed
 - FF-1429 - Updated Roslynator.Analyzers to 4.2.0
 - FF-1429 - Updated xunit.analyzers to 1.1.0
-- FF-3881 - Updated DotNet SDK to 7.0.101
 - FF-1429 - Updated Microsoft.NET.Test.Sdk to 17.4.1
 - FF-1429 - Updated SonarAnalyzer.CSharp to 8.51.0.59060
 - FF-1429 - Updated NonBlocking to 2.1.1
@@ -21,6 +20,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - FF-1429 - Updated Philips.CodeAnalysis.MaintainabilityAnalyzers to 1.2.32
 - FF-1429 - Updated Microsoft.Extensions to 7.0.2
 - FF-1429 - Updated Meziantou.Analyzer to 2.0.5
+- FF-3881 - Updated DotNet SDK to 7.0.102
 ### Removed
 ### Deployment Changes