+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - FF-1429 - Updated Meziantou.Analyzer to 2.0.13
+- FF-1429 - Updated FunFair.Test.Common to 6.0.18.2581
 ### Removed
 ### Deployment Changes