+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - Dependencies - Updated FunFair.Test.Common to 6.0.26.2754
 - Dependencies - Updated Microsoft.Extensions to 7.0.4
 - Dependencies - Updated Microsoft.NET.Test.Sdk to 17.5.0
+- Dependencies - Updated NuGet to 6.5.0
 ### Removed
 ### Deployment Changes