+++ CHANGELOG.md
@@ -13,12 +13,12 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - Dependencies - Updated Credfeto.Enumeration.Source.Generation to 1.0.7.19
-- Dependencies - Updated Meziantou.Analyzer to 2.0.22
 - Dependencies - Updated FunFair.Test.Common to 6.0.26.2754
 - Dependencies - Updated Microsoft.Extensions to 7.0.4
 - Dependencies - Updated Microsoft.NET.Test.Sdk to 17.5.0
 - Dependencies - Updated NuGet to 6.5.0
 - Dependencies - Updated SonarAnalyzer.CSharp to 8.55.0.65544
+- Dependencies - Updated Meziantou.Analyzer to 2.0.23
 ### Removed
 ### Deployment Changes