+++ CHANGELOG.md
@@ -11,10 +11,10 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Changed
 - SDK - Updated DotNet SDK to 7.0.203
 - Dependencies - Updated SonarAnalyzer.CSharp to 8.56.0.67649
-- Dependencies - Updated Meziantou.Analyzer to 2.0.36
 - Dependencies - Updated Credfeto.Enumeration.Source.Generation to 1.0.9.588
 - Dependencies - Updated FunFair.Test.Common to 6.0.30.633
 - Dependencies - Updated Roslynator.Analyzers to 4.3.0
+- Dependencies - Updated Meziantou.Analyzer to 2.0.37
 ### Removed
 ### Deployment Changes