+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - Added OptimizationPreference Policy
 - Dependencies - Updated Meziantou.Analyzer to 2.0.52
 - Dependencies - Updated Microsoft.VisualStudio.Threading.Analyzers to 17.6.40
+- Dependencies - Updated Microsoft.NET.Test.Sdk to 17.6.0
 ### Removed
 ### Deployment Changes