+++ CHANGELOG.md
@@ -15,6 +15,10 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 <!--
 Releases that have at least been deployed to staging, BUT NOT necessarily released to live.  Changes should be moved from [Unreleased] into here as they are merged into the appropriate release branch
 -->
+## [474.0.3] - 2023-06-20
+### Removed
+- Dotnet 8 preview
+
 ## [474.0.2] - 2023-06-02
 ### Changed
 - SDK - Updated DotNet SDK to 7.0.203